requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
duckdb>=0.10.0
httpx>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
jellyfish>=1.0.3
